requirements.txt
beautifulsoup4==4.12.2
mysql_connector_repackaged==0.3.1
pandas==2.0.3
Requests==2.31.0
